perf: simplify games tabs & pages
