+++ src/app/[locale]/games/dlcs/page.tsx
