+++ src/app/[locale]/games/series/page.tsx
