+++ src/components/dashboard/DashboardAppProvider.tsx
@@ -5,13 +5,16 @@ import { AppRouterCacheProvider } from '@mui/material-nextjs/v15-appRouter';
 // Hooks
 import {useTranslations} from 'next-intl';
 
-// icons
+// Icons
 import SportsEsportsIcon from '@mui/icons-material/SportsEsports';
 import ScheduleIcon from '@mui/icons-material/Schedule';
 import ScienceIcon from '@mui/icons-material/Science';
 import QueryStatsIcon from '@mui/icons-material/QueryStats';
 import HourglassEmptyIcon from '@mui/icons-material/HourglassEmpty';
 import LinkIcon from '@mui/icons-material/Link';
+import AppsIcon from '@mui/icons-material/Apps';
+import ListIcon from '@mui/icons-material/List';
+import ExtensionIcon from '@mui/icons-material/Extension';
 
 // Types
 import type { Navigation } from '@toolpad/core/AppProvider';
@@ -30,32 +33,48 @@ export default function DashboardAppProvider({
         {
             icon: <SportsEsportsIcon />,
             title: t("gamesKey"),
-            segment: "/games"
+            segment: "games",
+            children: [
+                {
+                    icon: <AppsIcon />,
+                    title: t("gamesTabs.grid")
+                },
+                {
+                    segment: "series",
+                    icon: <ListIcon />,
+                    title: t("gamesTabs.list")
+                },
+                {
+                    segment: "dlcs",
+                    icon: <ExtensionIcon />,
+                    title: t("gamesTabs.dlc")
+                }
+            ]
         },
         {
             icon: <ScheduleIcon />,
             title: t("planningKey"),
-            segment: "/planning"
+            segment: "planning"
         },
         {
             icon: <HourglassEmptyIcon />,
             title: t("backlog"),
-            segment: "/backlog"
+            segment: "backlog"
         },
         {
             icon: <ScienceIcon />,
             title: t("testsKey"),
-            segment: "/tests"
+            segment: "tests"
         },
         {
             icon: <QueryStatsIcon />,
             title: t("stats"),
-            segment: "/stats"
+            segment: "stats"
         },
         {
             icon: <LinkIcon />,
             title: t("links"),
-            segment: "/links"
+            segment: "links"
         },
     ];
